[1024] Scripts/Icecrown Citadel: Fixed players being teleported to entrance when Professor Putricide entered combat during Festergut or Rotface fight (encounter mechanic)
Core/Quests: Resolve ambiguity between QUEST_STATUS_COMPLETED and rewarded state by implementing QUEST_STATUS_REWARDED. The latter is not used in the database explicitly, only in the core. This is to prevent overhead by iterating over the rewarded quests storage multiple times.
All relevant upper level code has been updated accordingly to the best of my ability.

Scripts/Borean Tundra: Add Support for areatrigger conditions for Quest: Breaking Through (11898). Thanks to Svannon for original code and joint brainstorming session.

Core/Maps: Unload transports in MapManager::UnloadAll instead of singleton destructor.

Core/Transports: Properly delink units from transports on transport desutruction. - Thanks to Shauren for help.
Core/ObjectAccessor: Unload corpse 'storage' in added UnloadAll method called in WorldRunnable postservice hook.
Core/Master: Change some postservice queries to syncrhonous (direct) execution to ensure execution on shutdown
Core/Master: Remove redundant calls to ACE::init and ACE::fini

These changes were made to fix crash/freeze issues on shutdown.

patch by shauren & Machiavelli & click & leak.

+++ src/server/game/Achievements/AchievementMgr.cpp
@@ -583,14 +583,14 @@ void AchievementMgr::LoadFromDB(PreparedQueryResult achievementResult, PreparedQ
         {
             Field* fields = achievementResult->Fetch();
 
-            uint32 achievement_id = fields[0].GetUInt32();
+            uint32 achievement_id = fields[0].GetUInt16();
 
             // don't must happen: cleanup at server startup in sAchievementMgr->LoadCompletedAchievements()
             if (!sAchievementStore.LookupEntry(achievement_id))
                 continue;
 
             CompletedAchievementData& ca = m_completedAchievements[achievement_id];
-            ca.date = time_t(fields[1].GetUInt64());
+            ca.date = time_t(fields[1].GetUInt32());
             ca.changed = false;
             
             if (AchievementEntry const* pAchievement = sAchievementStore.LookupEntry(achievement_id))
@@ -605,9 +605,9 @@ void AchievementMgr::LoadFromDB(PreparedQueryResult achievementResult, PreparedQ
         {
             Field* fields = criteriaResult->Fetch();
 
-            uint32 id      = fields[0].GetUInt32();
+            uint32 id      = fields[0].GetUInt16();
             uint32 counter = fields[1].GetUInt32();
-            time_t date    = time_t(fields[2].GetUInt64());
+            time_t date    = time_t(fields[2].GetUInt32());
 
             AchievementCriteriaEntry const* criteria = sAchievementCriteriaStore.LookupEntry(id);
             if (!criteria)
